Add files via upload
Title of Invention:
Online gaming platform for electronic versions of board games with changeable digital in-game assets
ABSTRACT
A system and method for creating and playing online videogame versions of board games which contain identifiable digital in-game assets, for example games of chess family, to enrich the gaming experience, further deepen the gaming theory through bringing new diverse categories of players and systemically analyzing data from interrelated games and provide economic income for the plentitude of creators of digital in-game assets. The digital in-game assets may be player-tokens, characters, items, environments, special visual and audio effects, modes of play, GUIs, and other programming products. Digital in-game assets are created by independent creators on the electronic online gaming platform in accordance with rules and protocols of the platform, by the computational means and tools provided by the platform to be negotiated in the marketplace of the platform and to be transacted and further imported into the gameplays of the online videogames provided by the platform.
TECHNICAL FIELD
The present technology pertains to videogaming, and to creative digitizing those tabletop board games which have simple mechanics and physics and do not require vast computational resource in the electronic embodiments of gameplays. More specifically, the present technology may provide various techniques for deepening and individualizing gaming experience through alternating digital in-game assets and ludemes of the various versions of board games, especially so for the games of chess family.